% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_to_ms_name.R
\name{ms_language_to_ms_name}
\alias{ms_language_to_ms_name}
\alias{ms_validate_language_gender}
\title{Language and Gender to Microsoft Voice Name}
\usage{
ms_language_to_ms_name(language = "en-US", gender = c("Female", "Male"))

ms_validate_language_gender(language = "en-US", gender = c("Female", "Male"))
}
\arguments{
\item{language}{A language code, see \code{\link{ms_language_codes}}}

\item{gender}{Either Male or Female, not all languages support both genders}
}
\value{
A character string of the name of the voice
}
\description{
Language and Gender to Microsoft Voice Name
}
\examples{
ms_language_to_ms_name()
ms_validate_language_gender()
}
