% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_list_voices.R
\name{ms_list_voices}
\alias{ms_list_voices}
\title{List Voices}
\usage{
ms_list_voices(token = NULL, api_key = NULL, region = NULL, ...)
}
\arguments{
\item{token}{An authentication token, base-64 encoded usually from
\code{\link{ms_get_tts_token}}.  If not provided, will be created from
\code{\link{ms_get_tts_token}}}

\item{api_key}{Microsoft Cognitive Services API key, if token is not
provided.}

\item{region}{Subscription region for your key.
See \url{https://docs.microsoft.com/en-us/azure/cognitive-services/speech-service/overview}}

\item{...}{Additional arguments to send to \code{\link{GET}}}
}
\value{
A \code{data.frame} of the names and their long names.
}
\description{
List Voices
}
\examples{
\dontrun{
if (ms_have_tts_key()) {
res = ms_list_voices()
}

}

\dontrun{
res = ms_list_voices(region = "eastus")
}
}
