% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_locale_names.R
\docType{data}
\name{ms_locale_names}
\alias{ms_locale_names}
\title{Names of Microsoft Locales}
\format{
A \code{data.frame} with 2 columns:
\describe{
\item{code}{the 5 character langugae code, separated by a hyphen, also referred to as a locale}
\item{language}{the name of the language}
}
}
\usage{
ms_locale_names
}
\description{
Names of Microsoft Locales
}
\keyword{datasets}
