% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_synthesize.R
\name{ms_read_synthesis}
\alias{ms_read_synthesis}
\title{Read Synthesized output}
\usage{
ms_read_synthesis(output)
}
\arguments{
\item{output}{List from \code{\link{ms_synthesize}} with elements
\code{output_format} and \code{content}}
}
\value{
A Wave Object
}
\description{
Read Synthesized output
}
\note{
The \code{tuneR} package cannot read all different types of
the output here.
}
